Use double precision

+++ jubatus/client/anomaly_client.hpp
@@ -38,16 +38,16 @@ class anomaly : public jubatus::client::common::client {
     return f.get<std::vector<std::string> >();
   }
 
-  float update(const std::string& id,
+  double update(const std::string& id,
       const jubatus::client::common::datum& row) {
     msgpack::rpc::future f = c_.call("update", name_, id, row);
-    return f.get<float>();
+    return f.get<double>();
   }
 
-  float overwrite(const std::string& id,
+  double overwrite(const std::string& id,
       const jubatus::client::common::datum& row) {
     msgpack::rpc::future f = c_.call("overwrite", name_, id, row);
-    return f.get<float>();
+    return f.get<double>();
   }
 
   bool clear() {
@@ -55,9 +55,9 @@ class anomaly : public jubatus::client::common::client {
     return f.get<bool>();
   }
 
-  float calc_score(const jubatus::client::common::datum& row) {
+  double calc_score(const jubatus::client::common::datum& row) {
     msgpack::rpc::future f = c_.call("calc_score", name_, row);
-    return f.get<float>();
+    return f.get<double>();
   }
 
   std::vector<std::string> get_all_rows() {
